standardize calculation of dogtag cost
